+++ plugins/tasks/ansible/src/main/java/com/walmartlabs/concord/plugins/ansible/OutVarsProcessor.java
@@ -21,7 +21,6 @@ package com.walmartlabs.concord.plugins.ansible;
  */
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.sdk.Context;
 
 import java.io.IOException;
@@ -35,12 +34,13 @@ public class OutVarsProcessor {
 
     private Path outVarsFile;
 
-    public void prepare(Map<String, Object> args, Map<String, String> env, Path workDir, Path tmpDir) throws IOException {
+    public void prepare(Map<String, Object> args, Map<String, String> env, Path workDir, Path tmpDir) {
         String outVars = ArgUtils.getListAsString(args, AnsibleConstants.OUT_VARS_KEY);
         if (outVars == null) {
             return;
         }
         env.put("CONCORD_OUT_VARS", outVars);
+
         outVarsFile = tmpDir.resolve("out_vars.json");
         env.put("CONCORD_OUT_VARS_FILE", workDir.relativize(outVarsFile).toString());
     }