+++ plugins/tasks/ansible/src/main/resources/com/walmartlabs/concord/plugins/ansible/callback/concord_trace.py
@@ -2,6 +2,7 @@ import ujson as json
 import os
 import errno
 from ansible.plugins.callback import CallbackBase
+from concord_ansible_stats import ConcordAnsibleStats
 
 try:
     from __main__ import cli
@@ -39,31 +40,4 @@ class CallbackModule(CallbackBase):
         print "Trace saved to:", target_filename
 
     def playbook_on_stats(self, stats):
-        failures = stats.failures.keys()
-
-        unreachable = [e for e in stats.dark.keys()
-                       if e not in failures]
-
-        changed = [e for e in stats.changed.keys()
-                   if (e not in failures
-                       and e not in unreachable)]
-
-        ok = [e for e in stats.ok.keys()
-              if (e not in failures
-                  and e not in unreachable
-                  and e not in changed)]
-
-        skipped = [e for e in stats.skipped.keys()
-                   if (e not in failures
-                       and e not in unreachable
-                       and e not in changed
-                       and e not in ok)]
-
-        data = {
-            'ok': ok,
-            'failures': failures,
-            'unreachable': unreachable,
-            'changed': changed,
-            'skipped': skipped
-        }
-        self.log(data)
+        self.log(ConcordAnsibleStats.build_stats_data(stats))